反馈问题修改更新

+++ src/modules/admin/system/links/entities/links.entity.ts
@@ -32,7 +32,7 @@ export class Link {
   @Column('int', { name: 'price', comment: '商品价格' })
   price: number;
 
-  @Column('int', {
+  @Column('float', {
     name: 'coin',
     comment: '金币',
     nullable: true,
